+++ cpp/src/fftw/fftw_serial.cpp
@@ -12,33 +12,39 @@ class FFTW_C2C : public FFTBase {
     void* plan_forward_; void* plan_backward_;
 
 public:
-    FFTW_C2C(int ndim, const std::vector<int>& shape, py::array in, py::array out, std::string dtype)
+    FFTW_C2C(int ndim, const std::vector<int>& shape, py::array in, py::array out,
+             std::string dtype, int n_threads, unsigned flags)
         : shape_(shape), ndim_(ndim), N_(1), dtype_(dtype), plan_forward_(nullptr), plan_backward_(nullptr)
     {
         for (int i : shape_) N_ *= i;
         void* i_p = in.mutable_data(); void* o_p = out.mutable_data();
 
+#ifdef ENABLE_OPENMP
+        fftw_plan_with_nthreads(n_threads);
+        fftwf_plan_with_nthreads(n_threads);
+#endif
+
         if (dtype_ == "complex128") {
             if (ndim_ == 1) {
-                plan_forward_ = fftw_plan_dft_1d(shape_[0], (fftw_complex*)i_p, (fftw_complex*)o_p, FFTW_FORWARD, FFTW_ESTIMATE);
-                plan_backward_ = fftw_plan_dft_1d(shape_[0], (fftw_complex*)i_p, (fftw_complex*)o_p, FFTW_BACKWARD, FFTW_ESTIMATE);
+                plan_forward_ = fftw_plan_dft_1d(shape_[0], (fftw_complex*)i_p, (fftw_complex*)o_p, FFTW_FORWARD, flags);
+                plan_backward_ = fftw_plan_dft_1d(shape_[0], (fftw_complex*)i_p, (fftw_complex*)o_p, FFTW_BACKWARD, flags);
             } else if (ndim_ == 2) {
-                plan_forward_ = fftw_plan_dft_2d(shape_[0], shape_[1], (fftw_complex*)i_p, (fftw_complex*)o_p, FFTW_FORWARD, FFTW_ESTIMATE);
-                plan_backward_ = fftw_plan_dft_2d(shape_[0], shape_[1], (fftw_complex*)i_p, (fftw_complex*)o_p, FFTW_BACKWARD, FFTW_ESTIMATE);
+                plan_forward_ = fftw_plan_dft_2d(shape_[0], shape_[1], (fftw_complex*)i_p, (fftw_complex*)o_p, FFTW_FORWARD, flags);
+                plan_backward_ = fftw_plan_dft_2d(shape_[0], shape_[1], (fftw_complex*)i_p, (fftw_complex*)o_p, FFTW_BACKWARD, flags);
             } else if (ndim_ == 3) {
-                plan_forward_ = fftw_plan_dft_3d(shape_[0], shape_[1], shape_[2], (fftw_complex*)i_p, (fftw_complex*)o_p, FFTW_FORWARD, FFTW_ESTIMATE);
-                plan_backward_ = fftw_plan_dft_3d(shape_[0], shape_[1], shape_[2], (fftw_complex*)i_p, (fftw_complex*)o_p, FFTW_BACKWARD, FFTW_ESTIMATE);
+                plan_forward_ = fftw_plan_dft_3d(shape_[0], shape_[1], shape_[2], (fftw_complex*)i_p, (fftw_complex*)o_p, FFTW_FORWARD, flags);
+                plan_backward_ = fftw_plan_dft_3d(shape_[0], shape_[1], shape_[2], (fftw_complex*)i_p, (fftw_complex*)o_p, FFTW_BACKWARD, flags);
             }
         } else {
             if (ndim_ == 1) {
-                plan_forward_ = fftwf_plan_dft_1d(shape_[0], (fftwf_complex*)i_p, (fftwf_complex*)o_p, FFTW_FORWARD, FFTW_ESTIMATE);
-                plan_backward_ = fftwf_plan_dft_1d(shape_[0], (fftwf_complex*)i_p, (fftwf_complex*)o_p, FFTW_BACKWARD, FFTW_ESTIMATE);
+                plan_forward_ = fftwf_plan_dft_1d(shape_[0], (fftwf_complex*)i_p, (fftwf_complex*)o_p, FFTW_FORWARD, flags);
+                plan_backward_ = fftwf_plan_dft_1d(shape_[0], (fftwf_complex*)i_p, (fftwf_complex*)o_p, FFTW_BACKWARD, flags);
             } else if (ndim_ == 2) {
-                plan_forward_ = fftwf_plan_dft_2d(shape_[0], shape_[1], (fftwf_complex*)i_p, (fftwf_complex*)o_p, FFTW_FORWARD, FFTW_ESTIMATE);
-                plan_backward_ = fftwf_plan_dft_2d(shape_[0], shape_[1], (fftwf_complex*)i_p, (fftwf_complex*)o_p, FFTW_BACKWARD, FFTW_ESTIMATE);
+                plan_forward_ = fftwf_plan_dft_2d(shape_[0], shape_[1], (fftwf_complex*)i_p, (fftwf_complex*)o_p, FFTW_FORWARD, flags);
+                plan_backward_ = fftwf_plan_dft_2d(shape_[0], shape_[1], (fftwf_complex*)i_p, (fftwf_complex*)o_p, FFTW_BACKWARD, flags);
             } else if (ndim_ == 3) {
-                plan_forward_ = fftwf_plan_dft_3d(shape_[0], shape_[1], shape_[2], (fftwf_complex*)i_p, (fftwf_complex*)o_p, FFTW_FORWARD, FFTW_ESTIMATE);
-                plan_backward_ = fftwf_plan_dft_3d(shape_[0], shape_[1], shape_[2], (fftwf_complex*)i_p, (fftwf_complex*)o_p, FFTW_BACKWARD, FFTW_ESTIMATE);
+                plan_forward_ = fftwf_plan_dft_3d(shape_[0], shape_[1], shape_[2], (fftwf_complex*)i_p, (fftwf_complex*)o_p, FFTW_FORWARD, flags);
+                plan_backward_ = fftwf_plan_dft_3d(shape_[0], shape_[1], shape_[2], (fftwf_complex*)i_p, (fftwf_complex*)o_p, FFTW_BACKWARD, flags);
             }
         }
     }
@@ -87,33 +93,41 @@ public:
 class FFTW_R2C_OutPlace : public FFTBase {
     std::vector<int> shape_; int ndim_; ssize_t N_; std::string dtype_;
     void* plan_r2c_; void* plan_c2r_;
+
 public:
-    FFTW_R2C_OutPlace(int ndim, const std::vector<int>& shape, py::array r, py::array c, std::string dtype)
+    FFTW_R2C_OutPlace(int ndim, const std::vector<int>& shape, py::array r, py::array c,
+                      std::string dtype, int n_threads, unsigned flags)
         : shape_(shape), ndim_(ndim), N_(1), dtype_(dtype), plan_r2c_(nullptr), plan_c2r_(nullptr)
     {
         for(int i:shape_) N_*=i;
         void* rp = r.mutable_data(); void* cp = c.mutable_data();
+
+#ifdef ENABLE_OPENMP
+        fftw_plan_with_nthreads(n_threads);
+        fftwf_plan_with_nthreads(n_threads);
+#endif
+
         if (dtype_ == "float64") {
             if (ndim_==1) {
-                plan_r2c_ = fftw_plan_dft_r2c_1d(shape_[0], (double*)rp, (fftw_complex*)cp, FFTW_ESTIMATE);
-                plan_c2r_ = fftw_plan_dft_c2r_1d(shape_[0], (fftw_complex*)cp, (double*)rp, FFTW_ESTIMATE);
+                plan_r2c_ = fftw_plan_dft_r2c_1d(shape_[0], (double*)rp, (fftw_complex*)cp, flags);
+                plan_c2r_ = fftw_plan_dft_c2r_1d(shape_[0], (fftw_complex*)cp, (double*)rp, flags);
             } else if (ndim_==2) {
-                plan_r2c_ = fftw_plan_dft_r2c_2d(shape_[0],shape_[1], (double*)rp, (fftw_complex*)cp, FFTW_ESTIMATE);
-                plan_c2r_ = fftw_plan_dft_c2r_2d(shape_[0],shape_[1], (fftw_complex*)cp, (double*)rp, FFTW_ESTIMATE);
+                plan_r2c_ = fftw_plan_dft_r2c_2d(shape_[0],shape_[1], (double*)rp, (fftw_complex*)cp, flags);
+                plan_c2r_ = fftw_plan_dft_c2r_2d(shape_[0],shape_[1], (fftw_complex*)cp, (double*)rp, flags);
             } else {
-                plan_r2c_ = fftw_plan_dft_r2c_3d(shape_[0],shape_[1],shape_[2], (double*)rp, (fftw_complex*)cp, FFTW_ESTIMATE);
-                plan_c2r_ = fftw_plan_dft_c2r_3d(shape_[0],shape_[1],shape_[2], (fftw_complex*)cp, (double*)rp, FFTW_ESTIMATE);
+                plan_r2c_ = fftw_plan_dft_r2c_3d(shape_[0],shape_[1],shape_[2], (double*)rp, (fftw_complex*)cp, flags);
+                plan_c2r_ = fftw_plan_dft_c2r_3d(shape_[0],shape_[1],shape_[2], (fftw_complex*)cp, (double*)rp, flags);
             }
         } else {
             if (ndim_==1) {
-                plan_r2c_ = fftwf_plan_dft_r2c_1d(shape_[0], (float*)rp, (fftwf_complex*)cp, FFTW_ESTIMATE);
-                plan_c2r_ = fftwf_plan_dft_c2r_1d(shape_[0], (fftwf_complex*)cp, (float*)rp, FFTW_ESTIMATE);
+                plan_r2c_ = fftwf_plan_dft_r2c_1d(shape_[0], (float*)rp, (fftwf_complex*)cp, flags);
+                plan_c2r_ = fftwf_plan_dft_c2r_1d(shape_[0], (fftwf_complex*)cp, (float*)rp, flags);
             } else if (ndim_==2) {
-                plan_r2c_ = fftwf_plan_dft_r2c_2d(shape_[0],shape_[1], (float*)rp, (fftwf_complex*)cp, FFTW_ESTIMATE);
-                plan_c2r_ = fftwf_plan_dft_c2r_2d(shape_[0],shape_[1], (fftwf_complex*)cp, (float*)rp, FFTW_ESTIMATE);
+                plan_r2c_ = fftwf_plan_dft_r2c_2d(shape_[0],shape_[1], (float*)rp, (fftwf_complex*)cp, flags);
+                plan_c2r_ = fftwf_plan_dft_c2r_2d(shape_[0],shape_[1], (fftwf_complex*)cp, (float*)rp, flags);
             } else {
-                plan_r2c_ = fftwf_plan_dft_r2c_3d(shape_[0],shape_[1],shape_[2], (float*)rp, (fftwf_complex*)cp, FFTW_ESTIMATE);
-                plan_c2r_ = fftwf_plan_dft_c2r_3d(shape_[0],shape_[1],shape_[2], (fftwf_complex*)cp, (float*)rp, FFTW_ESTIMATE);
+                plan_r2c_ = fftwf_plan_dft_r2c_3d(shape_[0],shape_[1],shape_[2], (float*)rp, (fftwf_complex*)cp, flags);
+                plan_c2r_ = fftwf_plan_dft_c2r_3d(shape_[0],shape_[1],shape_[2], (fftwf_complex*)cp, (float*)rp, flags);
             }
         }
     }
@@ -162,12 +176,20 @@ public:
 class FFTW_R2C_InPlace : public FFTBase {
     std::vector<int> shape_; int ndim_; ssize_t N_; std::string dtype_;
     void* plan_r2c_; void* plan_c2r_;
+
 public:
-    FFTW_R2C_InPlace(int ndim, const std::vector<int>& shape, py::array data, std::string dtype)
+    FFTW_R2C_InPlace(int ndim, const std::vector<int>& shape, py::array data,
+                     std::string dtype, int n_threads, unsigned flags)
         : shape_(shape), ndim_(ndim), N_(1), dtype_(dtype), plan_r2c_(nullptr), plan_c2r_(nullptr)
     {
         for(int i:shape_) N_*=i;
         void* ptr = data.mutable_data();
+
+#ifdef ENABLE_OPENMP
+        fftw_plan_with_nthreads(n_threads);
+        fftwf_plan_with_nthreads(n_threads);
+#endif
+
         if (dtype_ == "float64") {
             if (ndim_==1) {
                 plan_r2c_ = fftw_plan_dft_r2c_1d(shape_[0], (double*)ptr, (fftw_complex*)ptr, FFTW_ESTIMATE);
@@ -237,7 +259,7 @@ class FFTW_Guru_C2C : public FFTBase {
 
 public:
     FFTW_Guru_C2C(const std::vector<int>& shape, const std::vector<int>& axes,
-                  py::array in, py::array out, std::string dtype)
+                  py::array in, py::array out, std::string dtype, int n_threads, unsigned flags)
         : dtype_(dtype)
     {
         std::vector<int> s_in = get_array_strides(in);
@@ -261,12 +283,18 @@ public:
         }
 
         void* ip = in.mutable_data(); void* op = out.mutable_data();
+
+#ifdef ENABLE_OPENMP
+        fftw_plan_with_nthreads(n_threads);
+        fftwf_plan_with_nthreads(n_threads);
+#endif
+
         if(dtype_ == "complex128") {
-            plan_fwd_ = fftw_plan_guru_dft(dims.size(), dims.data(), howmany_dims.size(), howmany_dims.data(), (fftw_complex*)ip, (fftw_complex*)op, FFTW_FORWARD, FFTW_ESTIMATE);
-            plan_bwd_ = fftw_plan_guru_dft(dims.size(), dims.data(), howmany_dims.size(), howmany_dims.data(), (fftw_complex*)ip, (fftw_complex*)op, FFTW_BACKWARD, FFTW_ESTIMATE);
+            plan_fwd_ = fftw_plan_guru_dft(dims.size(), dims.data(), howmany_dims.size(), howmany_dims.data(), (fftw_complex*)ip, (fftw_complex*)op, FFTW_FORWARD, flags);
+            plan_bwd_ = fftw_plan_guru_dft(dims.size(), dims.data(), howmany_dims.size(), howmany_dims.data(), (fftw_complex*)ip, (fftw_complex*)op, FFTW_BACKWARD, flags);
         } else {
-            plan_fwd_ = fftwf_plan_guru_dft(dims.size(), (fftwf_iodim*)dims.data(), howmany_dims.size(), (fftwf_iodim*)howmany_dims.data(), (fftwf_complex*)ip, (fftwf_complex*)op, FFTW_FORWARD, FFTW_ESTIMATE);
-            plan_bwd_ = fftwf_plan_guru_dft(dims.size(), (fftwf_iodim*)dims.data(), howmany_dims.size(), (fftwf_iodim*)howmany_dims.data(), (fftwf_complex*)ip, (fftwf_complex*)op, FFTW_BACKWARD, FFTW_ESTIMATE);
+            plan_fwd_ = fftwf_plan_guru_dft(dims.size(), (fftwf_iodim*)dims.data(), howmany_dims.size(), (fftwf_iodim*)howmany_dims.data(), (fftwf_complex*)ip, (fftwf_complex*)op, FFTW_FORWARD, flags);
+            plan_bwd_ = fftwf_plan_guru_dft(dims.size(), (fftwf_iodim*)dims.data(), howmany_dims.size(), (fftwf_iodim*)howmany_dims.data(), (fftwf_complex*)ip, (fftwf_complex*)op, FFTW_BACKWARD, flags);
         }
     }
 
@@ -318,7 +346,7 @@ class FFTW_Guru_R2C : public FFTBase {
 
 public:
     FFTW_Guru_R2C(const std::vector<int>& shape, const std::vector<int>& axes,
-                  py::array in, py::array out, std::string dtype)
+                  py::array in, py::array out, std::string dtype, int n_threads, unsigned flags)
         : dtype_(dtype)
     {
         std::vector<int> s_in = get_array_strides(in);
@@ -348,12 +376,17 @@ public:
 
         void* rp = in.mutable_data(); void* cp = out.mutable_data();
 
+#ifdef ENABLE_OPENMP
+        fftw_plan_with_nthreads(n_threads);
+        fftwf_plan_with_nthreads(n_threads);
+#endif
+
         if(dtype_ == "float64") {
-            plan_r2c_ = fftw_plan_guru_dft_r2c(df.size(), df.data(), hf.size(), hf.data(), (double*)rp, (fftw_complex*)cp, FFTW_ESTIMATE);
-            plan_c2r_ = fftw_plan_guru_dft_c2r(db.size(), db.data(), hb.size(), hb.data(), (fftw_complex*)cp, (double*)rp, FFTW_ESTIMATE);
+            plan_r2c_ = fftw_plan_guru_dft_r2c(df.size(), df.data(), hf.size(), hf.data(), (double*)rp, (fftw_complex*)cp, flags);
+            plan_c2r_ = fftw_plan_guru_dft_c2r(db.size(), db.data(), hb.size(), hb.data(), (fftw_complex*)cp, (double*)rp, flags);
         } else {
-            plan_r2c_ = fftwf_plan_guru_dft_r2c(df.size(), (fftwf_iodim*)df.data(), hf.size(), (fftwf_iodim*)hf.data(), (float*)rp, (fftwf_complex*)cp, FFTW_ESTIMATE);
-            plan_c2r_ = fftwf_plan_guru_dft_c2r(db.size(), (fftwf_iodim*)db.data(), hb.size(), (fftwf_iodim*)hb.data(), (fftwf_complex*)cp, (float*)rp, FFTW_ESTIMATE);
+            plan_r2c_ = fftwf_plan_guru_dft_r2c(df.size(), (fftwf_iodim*)df.data(), hf.size(), (fftwf_iodim*)hf.data(), (float*)rp, (fftwf_complex*)cp, flags);
+            plan_c2r_ = fftwf_plan_guru_dft_c2r(db.size(), (fftwf_iodim*)db.data(), hb.size(), (fftwf_iodim*)hb.data(), (fftwf_complex*)cp, (float*)rp, flags);
         }
     }
 
@@ -401,21 +434,30 @@ public:
 FFTW_SERIAL::FFTW_SERIAL(int ndim, const std::vector<int>& shape,
                          const std::vector<int>& axes,
                          py::array real_in, py::array complex_out,
-                         const std::string& dtype)
+                         const std::string& dtype, int n_threads, unsigned flags)
 {
+#ifdef ENABLE_OPENMP
+    static bool initialized = false;
+    if (!initialized) {
+        fftw_init_threads();
+        fftwf_init_threads();
+        initialized = true;
+    }
+#endif
+
     if (!axes.empty()) {
         if (dtype == "complex128" || dtype == "complex64") {
-            impl_ = std::make_unique<FFTW_Guru_C2C>(shape, axes, real_in, complex_out, dtype);
+            impl_ = std::make_unique<FFTW_Guru_C2C>(shape, axes, real_in, complex_out, dtype, n_threads, flags);
         } else {
-            impl_ = std::make_unique<FFTW_Guru_R2C>(shape, axes, real_in, complex_out, dtype);
+            impl_ = std::make_unique<FFTW_Guru_R2C>(shape, axes, real_in, complex_out, dtype, n_threads, flags);
         }
     } else {
         if (dtype == "complex128" || dtype == "complex64") {
-            impl_ = std::make_unique<FFTW_C2C>(ndim, shape, real_in, complex_out, dtype);
+            impl_ = std::make_unique<FFTW_C2C>(ndim, shape, real_in, complex_out, dtype, n_threads, flags);
         } else if (real_in.ptr() == complex_out.ptr()) {
-            impl_ = std::make_unique<FFTW_R2C_InPlace>(ndim, shape, real_in, dtype);
+            impl_ = std::make_unique<FFTW_R2C_InPlace>(ndim, shape, real_in, dtype, n_threads, flags);
         } else {
-            impl_ = std::make_unique<FFTW_R2C_OutPlace>(ndim, shape, real_in, complex_out, dtype);
+            impl_ = std::make_unique<FFTW_R2C_OutPlace>(ndim, shape, real_in, complex_out, dtype, n_threads, flags);
         }
     }
 }